+++ raiden_contracts/tests/fixtures/contracts.py
@@ -3,7 +3,7 @@ import logging
 from web3.utils.events import get_event_data
 from eth_utils import is_address
 
-from raiden_contracts.tests.utils.transaction import check_succesful_tx
+from raiden_contracts.utils.transaction import check_succesful_tx
 from raiden_contracts.constants import (
     CONTRACT_TOKEN_NETWORK,
     CONTRACT_TOKEN_NETWORK_REGISTRY,
