+++ raiden_contracts/tests/utils/__init__.py
@@ -3,5 +3,4 @@
 from .constants import *
 from .mock import *
 from .contracts import *
-from .transaction import *
 from .channel import *
